Create App.js
Inside src/App.js, we use the useEffect hook to make the request only when the component mounts by using an empty array [] for the “watch” parameter. We store the result with the useState hook and will pass it to our portfolio components which we’ll build in the upcoming sections.